+++ MvRAdaptiveCards/Public/Application/Out-CardPreview.ps1
@@ -132,6 +132,8 @@ function Out-CardPreview {
                 $ExtensionsCss += "`n/* Extension: $Extension */`n" + $ExtensionCssContent
             }
         }
+        $ExtensionsCss = "<style type='text/css'>$ExtensionsCss</style>"
+
         $html = $ExecutionContext.InvokeCommand.ExpandString($html)
 
         $path = "$env:TEMP\AdaptiveCardDesigner.html"