+++ MvRAdaptiveCards/Public/Inputs/New-CardInputToggle.ps1
@@ -159,7 +159,7 @@ function New-CardInputToggle {
         $InputToggle.title = $Title
     }
 
-    if ($PSBoundParameters.ContainsKey('Value')) {
+    if ($Value) {
         $InputToggle.value = $Value
     }